Add support for a CLI
