+++ requirements.txt
@@ -1,2 +1,3 @@
 requests==2.32.3
 twine==6.1.0
+click>=8.1.0
